requirements.txt
Click
python-docx
selenium